+++ src/frontend/apps/impress/src/features/docs/doc-management/hooks/useTrans.tsx
@@ -6,10 +6,32 @@ export const useTrans = () => {
   const { t } = useTranslation();
 
   const translatedRoles = {
-    [Role.ADMIN]: t('Administrator'),
     [Role.READER]: t('Reader'),
-    [Role.OWNER]: t('Owner'),
     [Role.EDITOR]: t('Editor'),
+    [Role.ADMIN]: t('Administrator'),
+    [Role.OWNER]: t('Owner'),
+  };
+
+  const getNotAllowedMessage = (
+    canUpdate: boolean,
+    isLastOwner: boolean,
+    isOtherOwner: boolean,
+  ) => {
+    if (!canUpdate) {
+      return undefined;
+    }
+
+    if (isLastOwner) {
+      return t(
+        'You are the sole owner of this group, make another member the group owner before you can change your own role or be removed from your document.',
+      );
+    }
+
+    if (isOtherOwner) {
+      return t('You cannot update the role or remove other owner.');
+    }
+
+    return undefined;
   };
 
   return {
@@ -17,5 +39,7 @@ export const useTrans = () => {
       return translatedRoles[role];
     },
     untitledDocument: t('Untitled document'),
+    translatedRoles,
+    getNotAllowedMessage,
   };
 };