+++ src/frontend/apps/impress/src/utils/children.ts
@@ -1,0 +1,9 @@
+import { Children, ReactNode } from 'react';
+
+export const hasChildrens = (element: ReactNode): boolean => {
+  let hasChildren = false;
+  Children.forEach(element, (child: ReactNode) => {
+    hasChildren = hasChildren || !!child;
+  });
+  return hasChildren;
+};