+++ CHANGELOG.md
@@ -20,7 +20,7 @@ and this project adheres to
 - 💄(frontend) update doc summary ui #473
 - 🐛(frontend) fix doc grid button #478
 - ✨(backend) add server-to-server API endpoint to create documents #467
-
+- ✨(frontend) new share modal ui #489
 ## Changed
 - 💄(frontend) add filtering from left panel #475